agrosense_backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.7
django-extensions==3.2.1
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
psycopg2==2.9.5
pycodestyle==2.10.0
python-dotenv==1.0.0
pytz==2022.7.1
sentinels==1.0.0
sqlparse==0.4.3
tzdata==2022.7
